% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_combine.R
\name{fasta_combine}
\alias{fasta_combine}
\title{Combine a list of DNAStringSet objects into one}
\usage{
fasta_combine(fasta_list, verbose = TRUE)
}
\arguments{
\item{fasta_list}{list of DNAStringSet objects (e.g. output of fasta_read()
with multiple paths)}

\item{verbose}{logical; if TRUE prints combined sequence count (default TRUE)}
}
\value{
DNAStringSet
}
\description{
Combine a list of DNAStringSet objects into one
}
